Move dataStore to repository
