+++ src/domain/repositories/InstanceRepository.ts
@@ -1,3 +1,4 @@
+import { StorageRepository } from "./StorageRepository";
 import { FutureData } from "../entities/Future";
 import { User } from "../entities/User";
 import { AggregatedRepository } from "./AggregatedRepository";
@@ -8,6 +9,7 @@ export interface InstanceRepository {
     metadata: MetadataRepository;
     events: EventsRepository;
     aggregated: AggregatedRepository;
+    dataStore: StorageRepository;
 
     getBaseUrl(): string;
     getCurrentUser(): FutureData<User>;
